### 2020-01-15 #13 @baioc
[Re] A Multi-Functional Synthetic Gene Network
**Original article:** [A Multi-Functional Synthetic Gene Network: A Frequency Multiplier, Oscillator and Switch](https://doi.org/10.1371/journal.pone.0016140)

**PDF URL:** [[Re] A Multi-Functional Synthetic Gene Network](https://github.com/baioc/re-multif/blob/master/article.pdf)
**Metadata URL:** [metadata.yaml](https://github.com/baioc/re-multif/blob/master/doc/metadata.yaml)
**Code URL:** [GitHub repository](https://github.com/baioc/re-multif)

**Scientific domain:** Biological Computing / Synthetic Biology
**Programming language:** Octave
**Suggested editor:** C. Titus Brown, Konrad Hinsen
