### 2020-01-26 #13 @degoldschmidt
@khinsen Thank you for inviting me to review, but unfortunately I have to decline, because the topic is out of my range.
