### 2020-03-11 #13 @aaronshifman
@khinsen appologies for the delay. @baioc thank you for the replication of this interesting article. It would appear that you've successfully replicated most of the original implementation with the exception to mainly your figure 3 which you address in the paper.

# Code Comments
This will not be a full review as at this stage the code is insufficient for publication. It would seem that you've done an excellent job trying to make each file both stand-alone and fairly well documented. The level of similarity between each simulation is incredibly high and as such it is very difficult to see where things differ between simulations and additionally what is simulation specific and what is general (i.e. euler stepping).

I will try to point places where this could be easily improved.

Firstly parameters. There are some parameters that are constant across all simulations (as far as I can tell). For example mRNA degradation rates are 2.5e-3. This could be refactored into a common location (or config file if you should so choose). Other parameters have a common value except in some simulations when they change i.e. in reduced simulations these could take default values and then update where necissary.

Secondly is the euler stepping. When trying to understand what a simulation is doing it is challenging in that I have to parse code that after reading I understand to be "unimportant" as its just the FE algorithm. I understand that you have a few different kinds of simulations with repeated pulses or single pulses. However most of this could be refactored into a few functions and instead of reading

```
Algorithm step 1
Algorithm step 2
.
.
.
Alogrithm step n
```
I could read

```
solution = fe(parameters)
```

and everything becomes much clearer.

Lastly is the fact that if I want to run your code I have no ability to just run everything. I have to run each file individually, and then move the saved figure to a common location for your pdf. I would be nice is there was at least one script to run everything and if they all saved figures into a common location.

This is not an exercise in "writing good code" but an important step. As being able to abstract the simulation into the "science" and the "implementation" allows for easier understanding of the simulation and the methods behind it as well as "debugging" in the case of a differing simulation from the reference.

n.b. For stochastic simulations FE in inappropriate due to noise scaling. See Euler-Maruyama alogrithm for a stochastic adaptation. This could be in part the challenge with your noise scaling in the stochastic replication. The authors do not mention how they performed this in their supplemental. However if they did use ode45 the adaptive tolerances will bias noise towards a lower amplitude.

# Implementation Comments

It would seem that most figures were well replicated. You've addressed the differences in your figure 3/ their figure 7. However I've found a few other discrepancies

* In your replications of their figure 5a-c (your figures 4-7) they have a longer run time which better shows the damped oscillations decaying to a steady state.

* In  their figure 8, your figure 9/10 there are two issues. Firstly if I've understood correctly you've combined their A/B and C/D into two figures instead of four. This does make it slightly more challenging to compare. Additionally they trigger the pulse at 1.5e5s whereas you trigger the pulse at 1e5s it would be good to have this the same.

# Article Comments

Overall I found the paper well organized and easy to read. My only large comment is that to get model equations I have to refer to the original manuscript. These should be included in the replicated manuscript.

Some minor comments

* Since figures are slightly out of order please provide a reference (or more explicitly label) which figures map to to which in the original implementation.
* Why does figure 13 have a different seed than figure 12 - this should be explained. Additionally it does not match the implementation.
