### 2020-03-13 #13 @baioc
@aaronshifman thank you for the kind review.

As you have pointed out, our intention was to make each script a standalone experiment, but I do agree that a more maintainable approach would be to configure each simulation individually in a separate file. We shall also look into the Euler-Maruyama algorithm for our stochastic simulations and after refactoring it should be easier to adjust the mentioned discrepancies in our figures, as well as to provide a script to run all experiments and generate figures accordingly.

Regarding the article, do you suggest we include the model's full ODE system, the QSSA approximation equations, or both?
