### 2020-03-13 #13 @aaronshifman
My inclination would be to include both, however since the full system is quite large if there is not a good way to present the whole system and the QSSA approximation then I would suggest including the approximation and a description (in text) of the full model.
