### 2020-03-16 #13 @khinsen
Thanks @aaronshifman for your detailed review, and thanks @baioc for your constructive first reaction. And apologies to both of you for being late (I am close to declaring GitHub notification bankruptcy). Finally, I am aware that I still need to find a second reviewer!
