### 2020-06-03 6aa701f
Acknowledge reviewer suggestions
Code went through some major refactoring:
- Scripts that were previously standalone experiments are now configuration files for a unique experiment runner.
- There is now a Python script which drives all experiments and generates figures.
- README.md was improved with a better description on how to reproduce our work.
- We now have a dependency on Octave Forge's signal package.

The replication document was also updated:
- Included the QSSA equations in the text.
- Minor discrepancies in Figures were addressed.
- Our Figures have explicit mappings to the ones i
[…]
