+++ src/stochastic/stochastic_freqdiv.m
@@ -1,202 +1,0 @@
-% ============================== OCTAVE SETUP ==================================
-    clear;
-    format long;
-
-
-% ========================== MODEL REACTION RATES ==============================
-
-    % translation rates (1/s)
-    Kt_R1 = 6e-4;
-    Kt_R2 = 6e-4;
-    Kt_R3 = 6e-4;
-    Kt_R4 = 6e-4;
-
-    % mRNA degradation rates (1/s)
-    Kd_mR1P1 = 2.5e-3;
-    Kd_mR4P1 = 2.5e-3;
-    Kd_mR2P2 = 2.5e-3;
-    Kd_mR3P2 = 2.5e-3;
-    Kd_mR1P3 = 2.5e-3;
-    Kd_mR2P4 = 2.5e-3;
-    Kd_mR3P5 = 2.5e-3;
-    Kd_mR4P6 = 2.5e-3;
-
-    % protein degradation rates (1/s)
-    Kd_R1 = 4e-4;
-    Kd_R2 = 4e-4;
-    Kd_R3 = 4e-4;
-    Kd_R4 = 4e-4;
-
-    % maximum/unrepressed transcription rates (M/s)
-    Kb_P1 = 4e-10;
-    Kb_P2 = 4e-10;
-    Kb_P3 = 4e-10;
-    Kb_P4 = 4e-10;
-    Kb_P5 = 4e-10;
-    Kb_P6 = 4e-10;
-
-    % Hill coefficients (scalar)
-    Na_P1 = 1.3;
-    Na_P2 = 1.3;
-    Nr_R2P1 = 1.3;
-    Nr_R4P2 = 1.3;
-    Nr_R3P3 = 1.3;
-    Nr_R3P4 = 1.3;
-    Nr_R4P4 = 1.3;
-    Nr_R1P5 = 1.3;
-    Nr_R1P6 = 1.3;
-    Nr_R2P6 = 1.3;
-
-    % Hill activation constants (M)
-    Ka_P1 = 2e-8;
-    Ka_P2 = 2e-8;
-
-    % Hill repression constants (M)
-    Kr_R2P1 = 6e-10;
-    Kr_R4P2 = 6e-10;
-    Kr_R3P3 = 6e-10;
-    Kr_R3P4 = 6e-10;
-    Kr_R4P4 = 6e-10;
-    Kr_R1P5 = 6e-10;
-    Kr_R1P6 = 6e-10;
-    Kr_R2P6 = 6e-10;
-
-
-% ============================ INPUT PARAMETERS ================================
-
-    % time settings
-    dt = 60;                     % 1 minute timesteps
-    simulation = 0 : dt : 4.0e5; % simulation length control
-
-    % input vector
-    period = 0.9e5;
-    amplitude = 50e-9;
-    minimum = 6e-9;
-    Is = minimum .- (amplitude/2)*cos(simulation .* 2*pi/period) .+ amplitude/2;
-
-    % initial state concentrations (M)
-    R1 = R2 = 50e-9;
-    R3 = R4 = 0e-9;
-
-
-% =============================== SIMULATION ===================================
-
-    % preallocate arrays to store concentrations over time
-    R1s = zeros(1, length(simulation));
-    R2s = zeros(1, length(simulation));
-    R3s = zeros(1, length(simulation));
-    R4s = zeros(1, length(simulation));
-
-    % precompute loop invariants and reduced coefficients
-    a1 = Kt_R1 * Kb_P1 / Kd_mR1P1;
-    a2 = Kt_R1 * Kb_P3 / Kd_mR1P3;
[…]
