+++ src/run_experiment.m
@@ -1,0 +1,249 @@
+#!/usr/bin/octave -qf
+% usage: octave script.m path/to/network.mat path/to/experiment.mat [load/path]
+
+clear;
+format long;
+
+pkg load signal; % only used for findpeaks()
+
+% parameterize experiment
+params = argv();
+load(params{1}); % network reaction constants
+load(params{2}); % experiment configuration
+if length(params) > 2
+  addpath(params{3}); % this is mandatory when using the python driver script
+endif
+
+
+% set initial protein concentrations
+R = [R1 R2 R3 R4];
+
+% choose input signal shape (period T left unfixed)
+if strcmp(input_type, 'sine')
+  I = @(t, T) input_amplitude * (-cos(t * 2*pi/T) + 1) / 2 + input_dc_level;
+elseif strcmp(input_type, 'square')
+  I = @(t, T) square_wave(t, T, ...
+                          input_amplitude, input_dc_level, input_duty_cycle);
+elseif strcmp(input_type, 'constant')
+  I = @(t, T) input_dc_level;
+  input_period = Inf;
+else
+  error("Undefined input type: %s.", input_type);
+endif
+
+% define short-hand for Hill functions
+k_a = [k_a k_a k_a k_a k_a];
+k_r = [k_r k_r k_r k_r];
+Ha = @(X, k) hill_activation(X, N, k);
+Hr = @(X, k) hill_repression(X, N, k);
+
+% defaults to zero noise, this is overrided when using stochastic model
+noise = @(t, len) 0;
+
+% choose the network's internal model (input I and constant k_r left unfixed)
+if strcmp(model, 'QSSA')
+  alpha = k_tl * beta / delta_m;
+  gamma = k_tl * P_tc / delta_m;
+  model = @(R, t, I, k_r) qssa(R, t, I, ...
+                               Ha, Hr, k_a, k_r, ...
+                               alpha, gamma, delta_x);
+elseif strcmp(model, 'Full')
+  R = [R [0 0 0 0 0 0 0 0]]; % insert starting mRNAs into molecule vector
+  model = @(R, t, I, k_r) full_ode(R, t, I, ...
+                                   Ha, Hr, k_a, k_r, ...
+                                   k_tl, beta, P_tc, delta_x, delta_m);
+elseif strcmp(model, 'Stochastic')
+  assert(strcmp(experiment_class, 'single') || strcmp(experiment_class, 'switch'));
+
+  % pre-generate Gaussian noise
+  randn('seed', random_seed);
+  steps = ceil(simulation_total_time / simulation_step);
+  noise = zeros(steps, 12);
+  for j = 1 : size(noise, 2)
+    noise(:,j) = randn(steps, 1) * noise_scaling;
+  endfor
+  noise_fn = @(t, col) noise(round(t / simulation_step) + 1, col);
+
+  % set model constructor
+  model = @(R, t, I, k_r) stochastic(R, t, I, noise_fn, ...
+                                     Ha, Hr, k_a, k_r, ...
+                                     k_tl, beta, P_tc, delta_x, delta_m);
+else
+  error("Undefined model type: %s.", model);
+endif
+
+
+% data scaling for easier visualization
+timescale = 1e5; % 10^5 seconds
+molscale = 1e-9; % nM
+
+% peak-based period detection, uses the last period for more stability
+function T = period(Y, dt = 1)
+  [pks, idx] = findpeaks(Y);
+  if length(pks) < 2
+    T = 0;
+  else
+    T = (idx(end) - idx(end-1)) * dt;
+  endif
+endfunction
+
+
+% run the simulation defined by the current experiment's class
+if strcmp(experiment_class, 'single')
+  % fix period and model
[…]
