### 2020-06-03 #13 @baioc
Apologies for the delay.
The replication repository's master branch has been updated to acknowledge reviewer comments and I believe most issues have been addressed.

The exception being the numerical integration method used in stochastic experiments, which remain the same. This was done to in order to preserve the Chemical Langevin Equations as defined in the original article's [Supporting Information File S6](https://doi.org/10.1371/journal.pone.0016140.s006).

**EDIT**: just noticed Figure 8's "data1" label, this has been removed in the latest commit.
