### 2020-07-30 #13 @aaronshifman
Sorry for the delay @khinsen and @baioc the revisions slipped by me.

I'll try to get a review back early next week. Just a cursory look would suggest @baioc has done a good first pass at refactoring the codebase. I do not understand your comment about preserving the Langevin Equations. If done properly the choice of integration method will not change the "meaning" of the differential equation. Using a standard Forward Euler will in effect scale the noise amplitude. (See [here](https://brian2.readthedocs.io/en/stable/user/models.html#time-scaling-of-noise) for a hand-waving justification).

I'll upload a formal review ASAP.
