### 2020-07-30 9389c30
Make some small fixes
- Removed the "data1" legend from Figure 8.
- Now using Zenodo's Concept (instead of version-specific) DOI.
- Small tweaks to README and to the SBOL diagram.
