+++ experiments.py
@@ -9,8 +9,7 @@ original_dir = os.curdir
 try:
     os.mkdir('temp')
 except FileExistsError:
-    print("WARNING: 'temp/' already exists, its contents may be overwritten.",
-          file=sys.stdout)
+    print("WARNING: 'temp/' already exists, its contents may be overwritten.", file=sys.stdout)
 finally:
     os.chdir('temp/')
 
@@ -19,8 +18,7 @@ experiments = '../src/parameters/experiments/'
 
 for config in os.listdir(experiments):
     print("\nRunning experiment configuration", config, "...")
-    sim = ['octave', '../src/run_experiment.m',
-           network, experiments + config, '../src']
+    sim = ['octave', '../src/run_experiment.m', network, experiments + config, '../src']
     proc = subprocess.run(sim, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
     print(proc.stdout.decode())
     print(proc.stderr.decode(), file=sys.stderr)
