+++ doc/metadata.tex
@@ -1,7 +1,7 @@
 % DO NOT EDIT - automatically generated from metadata.yaml
 
 \def \codeURL{https://github.com/baioc/re-multif}
-\def \codeDOI{10.5281/zenodo.3545452}
+\def \codeDOI{10.5281/zenodo.3545451}
 \def \dataURL{}
 \def \dataDOI{}
 \def \editorNAME{}
