### 2021-04-07 #13 @baioc
Greetings,

I was recently reminded of this ongoing submission and since it is now a little more than 1 year old, I wonder if I should consider it as rejected (and close the issue) or keep it open in case this is how long it usually takes.

@rougier I hope you don't mind that I tag you to make sure someone gets notified.
