### 2021-04-28 #13 @khinsen
Two quick replies.. but negative.

In order to move on with this, I propose a slight deviation from the usual ReScience process, if @oliviaguest, @benoit-girard, and @rougier don't have any objections: I auto-proclaim myself a reviewer for this paper.

My knowledge of synthetic biology is certainly weak, but it's only the computational aspects that matter for this replication, and they are about numerically solving a system of ordinary differential equations, which is something I have plenty of experience with. And I don't mind exploring the current state of Octave at the occasion, a piece of software that I haven't used for about 15 years.
