### 2021-04-30 #13 @rougier
Actually I think we wrote that an editor can also be a reviewer such that it is not even a deviation.
