### 2021-05-04 #13 @khinsen
@baioc Here comes my review of your article and code, in the current version that has already been improved based on the first review by @aaronshifman.

The submitted article attempts to replicate all of the numerical simulations in the original paper, using slightly different numerical techniques (a different ODE solver), and a new implementation written in Octave. In addition to replicating the simulation of the original work based on the quasi-steady state approximation (QSSA), the authors also perform some simulations using the more elaborate initial system of ODEs. Presumably that was not (or less) feasible ten years ago, when the original study was published. The assessment of QSSA as part of the replication adds to the interest of the replication article.

Most but not all of the the replication results agree very well with the figures shown in the original study. The biggest difference is observable in the bifurcation time scales of Fig. 3. The authors note the difference but provide no explanation, which is understandable given that the original source code is not available. There are also differences in the results of the stochastic simulations (described in supplementary material to the original work), where the authors had to modify the parameters quoted in the original work in order to get reasonably similar results.

The Octave code provided by the authors is not very complicated and very readable. I was able to run it code following the instructions, using the current version of Octave (6.2.0, running under macOS) rather than the version 5.1.0 used by the authors. For the benefit of long-term reproducibility, the authors should also mention the version numbers of the two add-on packages ("control" and "signal") that were used. For my evaluation, I used control 3.2.0 and signal 1.4.1.

Installation of "control" issued the error message "warning: LFLAGS is deprecated and will be removed in a future version of Octave, use LDFLAGS instead", which suggests that something may break in the future.

The figures that I obtained on my computers look very similar to those in the submitted paper. The only visible differences concern font sizes and line styles.
