### 2021-05-05 #13 @baioc
Thank you @khinsen for your review.

W.r.t. the additional Octave packages: the versions which were used are mentioned in the README.md file (I reckon there is some rephrasing that could be done to make that information more clear). Do you mean we should mention it directly in the article?
