### 2021-05-06 #13 @khinsen
@baioc My preference would be to mention all version numbers in the article, in particular if there are only three of them.

But I did have another look at the README.md file as well, and noticed a discreancy: the article says Octave 5.1.0, the README says 5.2.0.

I also wonder if Octave actually provides a way to install specific versions of dependencies. I couldn't find any information on this, but I didn't spend that much time searching for it either. If it is possible and not too complicated, it would be nice to reformulate the instructions in README.md accordingly.
