### 2021-05-08 #13 @baioc
I believe the most recent commits to the [dev branch](https://github.com/baioc/re-multif/tree/dev) address these comments: package versions are now mentioned in the replication paper and install instructions in the README now fetch, from Octave Forge, the specific versions which were used.

I also re-ran all experiments in Octave 6.2.0 (under Manjaro-Linux) and can confirm they still work, with only minor stylistic changes to plots, as reported by @khinsen.
