### 2021-05-27 #13 @aaronshifman
Hello @baioc thank you for the updates to your codebase along with the manuscript.

I believe that the codebase is now more readable and packages the experiments nicely, and faithfully reproduces the original manuscript, with the exceptions that that authors address in Figure 3. As @khinsen mentioned the reference work was done a long time ago and certainly I'm sure tolerances due to lack of computation power would be quite lax. That said, it would be nice to implement this system in a bifurcation suite (XPPAUT or MATCONT) and confirm the new results (or re-running the simulation with a smaller time step).

My only hanging issue is the stochastic simulations. The use of the FE algorithm will incorrectly scale the noise by \sqrt(dt)... however since the original authors used the adaptive ode45 in MATLAB (also not an ideal method but for a different reason - noise down scaling during fast transients). I thought this replication could serve as an inference as to the effective noise-scale in the original implementation. However, since the noise scale in both cases are in a sense "incorrect", but in different ways. I would defer this decision to whether this should be corrected to the editor.

Other than these two comments, I feel like this article is suitable for publication.
