### 2021-05-31 #13 @khinsen
Thanks @aaronshifman for your comments!

From my point of view as a ReScience editor (rather than a domain specialist), the first proposition (improving bifurcation analysis) is certainly of interest but also well beyond doing a replication. If @baioc wants to do this as an added-value section, that's fine with me, but it's not a requirement.

The second point is more relevant from a replication point of view. The replication uses a different integration algorithm compared to the original work, and yields different noise patterns for the stochastic simulations. The minimal requirement is discussing this issue in the replication paper, which @baioc did. The best next step from a purely "replication" point of view would be to use ode45 for the stochastic simulations and see if the noise patterns are replicable. This could well fail because of different ode45 implementations in Matlab vs. Octave.

Beyond such a check, one can aim aim for insight into what is happening, which is what @aaronshifman suggests. As for his first point, this is a useful addition but not not a requirement for ReScience. We are dealing with suboptimal practices here: using ODE solvers for SDE is fundamentally wrong, but also common practice in many fields. My personal preference for shedding light on the impact of this mistake would be the use of a proper SDE integrator. But that could well turn out to become a research project of its own, depending on the state of the art in this particular type of equations, which I am not familiar with.

@baioc What's your point of view on this?
