### 2021-06-13 #13 @baioc
First I would like to thank @aaronshifman and @khinsen for the reviews, and apologize for the delay on the reply.
Since this has been going on for well over a year, I would prefer not to go over significant changes to the implementation and leave using an SDE aside since I'm not familiar with those at all.

That said, if running the stochastic simulations using Octave's ode45 doesn't prove being much of a problem, I will compare it to the FE method and see if I can reproduce the oscillator mode irregularities.
Also with respect to the noise scaling in FE, this is somewhat addressed by the downscaling mentioned in section 3.5, but I can definitely try to do a better job at explaining how that is caused by the integration method being used.

As for the bifurcation results, I will run some more simulations varying the timestep (which is something that may have been already done earlier last year, so I don't think the results are going to be different) and mention those in the text as well.

However, I am quite busy this month, so expect this to be done only by the end of July.
