### 2021-12-01 #13 @baioc
@khinsen apologies for being late. It turns out that my prediction of when I'd have the time to tackle this was way off (it also doesn't help that my studies and work have moved far from the subject of this paper).
Now that we're approaching the end of the year, however, I'll soon have some free time to work on this. I'll let you know once the repository is updated with the additional simulations and the accompanying commentary is added to the article text.
