### 2022-01-06 335213b
Update replication
- Added the option to use Octave's ode45 solver instead of our FE method.
- Updated the article text with comments on different integration methods and time step variations.
