+++ src/stochastic.m
@@ -2,11 +2,11 @@
 % Stochastic model) for given molecule concentrations R at instant t. I(t)
 % should evaluate to the input signal at time t; Ha(X,k) and Hr(X,k) must
 % compute the Hill functions for activation and repression of protein X given
-% half-saturation constant k; and k_tl beta, P_tc, delta_x and delta_m are
+% half-saturation constant k; and k_tl, beta, P_tc, delta_x and delta_m are
 % reaction constants extracted from the network. k_a and k_r are vectors with
 % the binding affinity for each repressor protein + input. N is used as a matrix
 % where each row represents noise at time t, with 12 columns for the coefficients.
-function dR = stochastic(R, t, I, N, Ha, Hr, k_a, k_r, k_tl, beta, P_tc, delta_x, delta_m)
+function dR = stochastic(t, R, I, N, Ha, Hr, k_a, k_r, k_tl, beta, P_tc, delta_x, delta_m)
 
   % these are just coefficient indexes for the noise matrix
   a1_idx = 1;
@@ -47,16 +47,12 @@ function dR = stochastic(R, t, I, N, Ha, Hr, k_a, k_r, k_tl, beta, P_tc, delta_x
 
   dR = zeros(size(R));
 
-  dR(1) = (sqrt(a1)*N(t,a1_idx) + sqrt(a2)*N(t,a2_idx) - sqrt(a3)*N(t,a3_idx)) ...
-          / sqrt(V) + a1 + a2 - a3;
+  dR(1) = a1 + a2 - a3 + (sqrt(a1)*N(t,a1_idx) + sqrt(a2)*N(t,a2_idx) - sqrt(a3)*N(t,a3_idx)) / sqrt(V);
 
-  dR(2) = (sqrt(b1)*N(t,b1_idx) + sqrt(b2)*N(t,b2_idx) - sqrt(b3)*N(t,b3_idx)) ...
-          / sqrt(V) + b1 + b2 - b3;
+  dR(2) = b1 + b2 - b3 + (sqrt(b1)*N(t,b1_idx) + sqrt(b2)*N(t,b2_idx) - sqrt(b3)*N(t,b3_idx)) / sqrt(V);
 
-  dR(3) = (sqrt(c1)*N(t,c1_idx) + sqrt(c2)*N(t,c2_idx) - sqrt(c3)*N(t,c3_idx)) ...
-          / sqrt(V) + c1 + c2 - c3;
+  dR(3) = c1 + c2 - c3 + (sqrt(c1)*N(t,c1_idx) + sqrt(c2)*N(t,c2_idx) - sqrt(c3)*N(t,c3_idx)) / sqrt(V);
 
-  dR(4) = (sqrt(d1)*N(t,d1_idx) + sqrt(d2)*N(t,d2_idx) - sqrt(d3)*N(t,d3_idx)) ...
-          / sqrt(V) + d1 + d2 - d3;
+  dR(4) = d1 + d2 - d3 + (sqrt(d1)*N(t,d1_idx) + sqrt(d2)*N(t,d2_idx) - sqrt(d3)*N(t,d3_idx)) / sqrt(V);
 
 endfunction
