+++ src/run_experiment.m
@@ -36,39 +36,42 @@ k_r = [k_r k_r k_r k_r];
 Ha = @(X, k) hill_activation(X, N, k);
 Hr = @(X, k) hill_repression(X, N, k);
 
-% defaults to zero noise, this is overrided when using stochastic model
-noise = @(t, len) 0;
-
 % choose the network's internal model (input I and constant k_r left unfixed)
 if strcmp(model, 'QSSA')
   alpha = k_tl * beta / delta_m;
   gamma = k_tl * P_tc / delta_m;
-  model = @(R, t, I, k_r) qssa(R, t, I, ...
+  model = @(t, R, I, k_r) qssa(t, R, I, ...
                                Ha, Hr, k_a, k_r, ...
                                alpha, gamma, delta_x);
 elseif strcmp(model, 'Full')
   R = [R [0 0 0 0 0 0 0 0]]; % insert starting mRNAs into molecule vector
-  model = @(R, t, I, k_r) full_ode(R, t, I, ...
+  model = @(t, R, I, k_r) full_ode(t, R, I, ...
                                    Ha, Hr, k_a, k_r, ...
                                    k_tl, beta, P_tc, delta_x, delta_m);
 elseif strcmp(model, 'Stochastic')
   assert(strcmp(experiment_class, 'single') || strcmp(experiment_class, 'switch'));
 
-  % pre-generate Gaussian noise
+  % prepare Gaussian noise
   randn('seed', random_seed);
-  steps = ceil(simulation_total_time / simulation_step);
-  noise = zeros(steps, 12);
-  for j = 1 : size(noise, 2)
-    noise(:,j) = randn(steps, 1) * noise_scaling;
-  endfor
-  noise_fn = @(t, col) noise(round(t / simulation_step) + 1, col);
+  noise_fn = @(t, col) randn() * noise_scaling;
 
   % set model constructor
-  model = @(R, t, I, k_r) stochastic(R, t, I, noise_fn, ...
+  model = @(t, R, I, k_r) stochastic(t, R, I, noise_fn, ...
                                      Ha, Hr, k_a, k_r, ...
                                      k_tl, beta, P_tc, delta_x, delta_m);
 else
   error("Undefined model type: %s.", model);
+endif
+
+% choose integration method to be used
+dt = simulation_step;
+if strcmp(method, 'Euler')
+  simulate = @(fun, interval, init) euler_simulate(fun, init, interval, dt);
+elseif strcmp(method, 'ode45')
+  odeopts = odeset("RelTol", 1e-10, "AbsTol", 1e-10, "InitialStep", dt);
+  simulate = @(fun, interval, init) ode45(fun, interval, init, odeopts);
+else
+  error("Undefined integration method: %s.", method);
 endif
 
 
@@ -77,12 +80,12 @@ timescale = 1e5; % 10^5 seconds
 molscale = 1e-9; % nM
 
 % peak-based period detection, uses the last period for more stability
-function T = period(Y, dt = 1)
+function T = period(Y, timeof = @(i) i)
   [pks, idx] = findpeaks(Y);
   if length(pks) < 2
     T = 0;
   else
-    T = (idx(end) - idx(end-1)) * dt;
+    T = timeof(idx(end)) - timeof(idx(end-1));
   endif
 endfunction
 
@@ -91,24 +94,22 @@ endfunction
 if strcmp(experiment_class, 'single')
   % fix period and model
   I = @(t) I(t, input_period);
-  model = @(R, t) model(R, t, I, k_r);
+  model = @(t, R) model(t, R, I, k_r);
 
   % run single simulatin
-  step_number = ceil(simulation_total_time / simulation_step);
-  Rs = euler_simulate(model, R, step_number, simulation_step);
+  [t, Rs] = simulate(model, [0.0, simulation_total_time], R);
[…]
