### 2022-01-06 a77619b
Update replication (#3)
- Added the option to use Octave's ode45 solver instead of our FE method.
- Updated the article text with comments on different integration methods and time step variations.
- Also merging previous changes that were only in the dev branch.
