+++ src/save_parameters.m
@@ -4,6 +4,7 @@
 % experimental context parameters (Stochastic only works w/ single || switch)
 experiment_class = 'single'; % 'single' | 'period' | 'oscillator' | 'switch'
 model = 'Stochastic'; % 'QSSA' | 'Full' | 'Stochastic'
+method = 'Euler'; % 'Euler' | 'ode45'
 
 % only used in the 'Stochastic' model
 random_seed =  0.8977812586558097;
@@ -48,7 +49,7 @@ plot_aspect = [1 0.334 1]; % vector or "auto" that is passed to pbaspect()
 
 % save experiment configuration in plain text
 save "parameters.mat" ...
-     experiment_class model ...
+     experiment_class model method ...
      random_seed noise_scaling ...
      period_experiment_range ...
      oscillator_input_range ...
