### 2022-01-06 #13 @baioc
Alright, I added an extra simulation configuration for the integration method to be used and was able to reproduce all experiments using Octave's ode45 solver. Results match those obtained using FE, including in the stochastic simulations (we still needed to compensate noise variance) and the same oscillator irregularities could be observed.

With respect to the period-doubling bifurcation, varying the Euler time step between 1 and 60 (or using the adaptive integrator) showed no significant difference that could explain the huge offset between the bifurcation's location in the reproduction and in the original work. The possible explanation for this remains the same as before: there could have been some mismatch in the parameters which were actually used in the original paper.

The article text was updated to take those additional experiments into consideration and the repository's main branch has been updated as well. Hopefully these changes address the remaining comments that came up in the review.
