### 2022-01-10 #13 @khinsen
Thanks @baioc for the updated article and code!  To me this looks very good.

@aaronshifman Any comments from your side?
