### 2022-07-02 #13 @aaronshifman
@khinsen Apologies for the delay, after looking at the revisions made by @baioc, I believe they've adequately addressed my concerns. I have no further issues with the state or the article or codebase.
