### 2022-07-05 #13 @khinsen
Thanks @aaronshifman !

So... I think we can... finally... proclaim this submission accepted!

I will handle the publishing step tomorrow. @baioc, I may well request help from you in producing the final manuscript, depending on how things turn out.
