+++ doc/article.py
@@ -167,7 +167,7 @@ class Article:
             
         
     def parse(self, data):
-        document = yaml.load(data)
+        document = yaml.load(data, Loader=yaml.CLoader)
 
         self.title = document.get("title", "")
         self.abstract = document.get("abstract","") or ""
