### 2022-07-06 #13 @khinsen
The article is now published:  https://zenodo.org/record/6801765/files/article.pdf

@baioc: Please accept pull request https://github.com/baioc/re-multif/pull/4, which adds the metadata added during publication.

Thanks to everyone, in particular @aaronshifman  for the review and @rougier for the regular reminders without which this would have taken even longer!
