### 2022-07-06 3ab84df
Post-publication commit
This work has been published in the ReScience journal!

This last commit updates DOI badges in the README and adds a special CITATION.cff file.
It also updates the generated metadata file and commits the published article.

Lastly, all figures were re-generated in Octave 7.1.0, which required a small fix in the experiment script.
It fixes the problem reported here: https://octave.discourse.group/t/plot-with-is-not-usable-on-7-1-0/2867
