+++ doc/metadata.tex
@@ -4,21 +4,21 @@
 \def \codeDOI{10.5281/zenodo.3545451}
 \def \dataURL{}
 \def \dataDOI{}
-\def \editorNAME{}
-\def \editorORCID{}
-\def \reviewerINAME{}
-\def \reviewerIORCID{}
-\def \reviewerIINAME{}
-\def \reviewerIIORCID{}
-\def \dateRECEIVED{}
-\def \dateACCEPTED{}
-\def \datePUBLISHED{}
+\def \editorNAME{Konrad Hinsen}
+\def \editorORCID{0000-0003-0330-9428}
+\def \reviewerINAME{Aaron Shifman}
+\def \reviewerIORCID{0000-0003-2140-7590}
+\def \reviewerIINAME{Konrad Hinsen}
+\def \reviewerIIORCID{0000-0003-0330-9428}
+\def \dateRECEIVED{15 January 2020}
+\def \dateACCEPTED{05 July 2022}
+\def \datePUBLISHED{06 July 2022}
 \def \articleTITLE{[Re] A Multi-Functional Synthetic Gene Network}
 \def \articleTYPE{Replication}
 \def \articleDOMAIN{Biological Computing}
 \def \articleBIBLIOGRAPHY{bibliography.bib}
 \def \articleYEAR{2022}
-\def \reviewURL{}
+\def \reviewURL{https://github.com/ReScience/submissions/issues/13}
 \def \articleABSTRACT{Studies in the field of synthetic biology are constantly making additions to biological circuit repositories, as well as to the theoretical understanding of their capabilities. The ability to compute with biomolecules has been demonstrated by many synthetic gene networks, but until recently the majority of such models were engineered to implement the functionality of a single circuit part. Purcell et al. have proposed a network capable of multiple functions, switching between three different behaviours in a programmable fashion. This work provides an open-source implementation in which their in silico experiments were replicated.}
 \def \replicationCITE{Purcell O, di Bernardo M, Grierson CS, Savery NJ (2011) A Multi-Functional Synthetic Gene Network: A Frequency Multiplier, Oscillator and Switch. PLOS ONE 6(2): e16140.}
 \def \replicationBIB{multif}
@@ -28,10 +28,10 @@
 \def \contactEMAIL{baiocchi.gabriel@gmail.com}
 \def \articleKEYWORDS{rescience c, octave, biological computing, synthetic biology, programmable genetic circuit, frequency multiplier, oscillator, toggle switch}
 \def \journalNAME{ReScience C}
-\def \journalVOLUME{4}
+\def \journalVOLUME{8}
 \def \journalISSUE{1}
-\def \articleNUMBER{}
-\def \articleDOI{}
+\def \articleNUMBER{5}
+\def \articleDOI{10.5281/zenodo.6801765}
 \def \authorsFULL{Gabriel Baiocchi de Sant'Anna and Mateus Favarin Costa}
 \def \authorsABBRV{G.B. Sant'Anna and M.F. Costa}
 \def \authorsSHORT{Sant'Anna and Costa}
