### 2022-07-06 #13 @baioc
Awesome! I've accepted the pull request and updated the Zenodo code archive.

@khinsen @aaronshifman @rougier I really appreciate your effort and patience throughout the review process.
Thank you all!
